Fusion avec la branche alphaseny pour la page d'inscription
