+++ front-end/src/App.jsx
@@ -21,14 +21,14 @@ import Details from './Composants/Details/Details.jsx';
 import Parametres from './Composants/Parametre/Parametres.jsx';
 import Connexion from './pages/Connexion/Connexion.jsx';
 import InfosPme from './pages/InfosPME/infosPME.jsx';
-import Register from './pages/Register/register.jsx';
+import Inscription from './pages/inscription/inscription.jsx';
 
 function App() {
   const location = useLocation();
 
   // Condition pour vérifier si on est sur une page qui ne doit pas afficher Navbar et Footer
   const hideNavbarAndFooter =
-    location.pathname === '/register' ||
+  location.pathname === '/inscription' ||
     location.pathname === '/connexion' ||  // Add this line to hide Navbar and Footer on the login page
     location.pathname.startsWith('/dashboard');
 
@@ -39,7 +39,7 @@ function App() {
 
       <Routes>
         <Route path='/' element={<Accueil />} />
-        <Route path='/register' element={<Register />} />
+        <Route path='/inscription' element={<Inscription />} />
         <Route path='/sens' element={<Sensibilisation />} />
         <Route path='/pmes' element={<Pmes />} />
         <Route path='/pmes/:id' element={<InfosPme/>}/>
